pyproject.toml
[tool.poetry]
name = "poetryproject"
version = "0.1.0"
description = ""
authors = ["evgen.arc <my@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest-cov = "^5.0.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
mypy = "^1.10.1"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119

[tool.isort]
line_length = 119

